Rename Statistics to HiveStatistics
