+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveMetadata.java
@@ -250,11 +250,11 @@ import static io.trino.plugin.hive.util.HiveWriteUtils.checkTableIsWritable;
 import static io.trino.plugin.hive.util.HiveWriteUtils.initializeSerializer;
 import static io.trino.plugin.hive.util.HiveWriteUtils.isS3FileSystem;
 import static io.trino.plugin.hive.util.HiveWriteUtils.isWritableType;
-import static io.trino.plugin.hive.util.Statistics.ReduceOperator.ADD;
-import static io.trino.plugin.hive.util.Statistics.createComputedStatisticsToPartitionMap;
-import static io.trino.plugin.hive.util.Statistics.createEmptyPartitionStatistics;
-import static io.trino.plugin.hive.util.Statistics.fromComputedStatistics;
-import static io.trino.plugin.hive.util.Statistics.reduce;
+import static io.trino.plugin.hive.util.HiveStatistics.ReduceOperator.ADD;
+import static io.trino.plugin.hive.util.HiveStatistics.createComputedStatisticsToPartitionMap;
+import static io.trino.plugin.hive.util.HiveStatistics.createEmptyPartitionStatistics;
+import static io.trino.plugin.hive.util.HiveStatistics.fromComputedStatistics;
+import static io.trino.plugin.hive.util.HiveStatistics.reduce;
 import static io.trino.plugin.hive.util.SystemTables.createSystemTable;
 import static io.trino.spi.StandardErrorCode.INVALID_ANALYZE_PROPERTY;
 import static io.trino.spi.StandardErrorCode.INVALID_SCHEMA_PROPERTY;
