+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/util/TestStatistics.java
@@ -37,11 +37,11 @@ import static io.trino.plugin.hive.HiveBasicStatistics.createZeroStatistics;
 import static io.trino.plugin.hive.metastore.HiveColumnStatistics.createBinaryColumnStatistics;
 import static io.trino.plugin.hive.metastore.HiveColumnStatistics.createBooleanColumnStatistics;
 import static io.trino.plugin.hive.metastore.HiveColumnStatistics.createIntegerColumnStatistics;
-import static io.trino.plugin.hive.util.Statistics.ReduceOperator.ADD;
-import static io.trino.plugin.hive.util.Statistics.ReduceOperator.SUBTRACT;
-import static io.trino.plugin.hive.util.Statistics.createHiveColumnStatistics;
-import static io.trino.plugin.hive.util.Statistics.merge;
-import static io.trino.plugin.hive.util.Statistics.reduce;
+import static io.trino.plugin.hive.util.HiveStatistics.ReduceOperator.ADD;
+import static io.trino.plugin.hive.util.HiveStatistics.ReduceOperator.SUBTRACT;
+import static io.trino.plugin.hive.util.HiveStatistics.createHiveColumnStatistics;
+import static io.trino.plugin.hive.util.HiveStatistics.merge;
+import static io.trino.plugin.hive.util.HiveStatistics.reduce;
 import static io.trino.spi.predicate.Utils.nativeValueToBlock;
 import static io.trino.spi.statistics.ColumnStatisticType.MAX_VALUE;
 import static io.trino.spi.statistics.ColumnStatisticType.MIN_VALUE;
